refactor: App의 상태 각 페이지로 이동
- 카드 추가 폼 페이지에서 추가된 카드 정보 history.push를 이용하여 카드 추가 완료 페이지로 전달
